Add Mitali's officer intro
The officer photos were added to the server but I don't feel that
it's necessary to add the photos to our GitHub server (ever) since
I'm pretty sure we could use the version control to find the
pictures of every generation of officers (if we continue to use
GitHub and uploaded officer photos). That's unnecessary and could
cause privacy issues later on, so personally I think it's best to
only store the photos on our server.

With this, the website is done! Huzzah!!