Changed file name, special arrow sign, and some code aesthetics
Changed the special arrow character to `<--` since the lab machines
(and a lot of older computers) don't render the fancy arrow symbol.
We found this out last time we changed the website, but we forgot to
change the code.

Other changes were just to preserve the spacing patterns in the code.
If we start adding a lot to the website, we should consider adding a
linter to run in GitHub actions. That way, we keep a consistent code
style without forcing people to install npm (we can't install anything
on the math lab machines, which is why we have such a bare-bones
website).

Awesome job, Aileen! This was great work and a super fast turnaround!